carmanager/requirements.txt
asgiref==3.8.1
Django==5.0.8
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.1
